Add missions to profile
