+++ src/App.css
@@ -3,46 +3,43 @@
   justify-content: space-between;
   font-weight: bold;
   display: grid;
-  grid-template-columns: 1fr 4fr 1fr 1.2fr;
+  grid-template-columns: 1fr 4fr 1fr 1fr;
   list-style: none;
-  margin: 0 30px;
+  margin: 20px 30px 0 30px;
   text-align: center;
 }
 
 .missions-info > * {
   padding: 5px;
-  border: solid 2px #d3d3d3;
+  border: 1px solid #d3d3d3;
 }
 
 .missions-details {
   display: grid;
-  grid-template-columns: 1fr 4fr 1fr 1.2fr;
-  border: solid 1.5px #d3d3d3;
+  grid-template-columns: 1fr 4fr 1fr 1fr;
+  border: 1px solid #d3d3d3;
   list-style: none;
   margin: 0 30px;
   text-align: center;
-  background-color: rgb(242, 242, 242);
-}
-
-.missions-details:nth-child(odd) {
-  background-color: white;
-}
-
-.missions-details > * {
-  padding: 10px;
-  border-right: solid 2px #d3d3d3;
 }
 
 .missions-details h3 {
   font-size: 16px;
+  padding: 10px;
+  border-right: solid 1px #d3d3d3;
+}
+
+.missions-details p {
+  padding: 10px;
+  border-right: solid 1px #d3d3d3;
 }
 
 .missions-details .notActive {
   background-color: grey;
   color: white;
-  border-radius: 3px;
+  border-radius: 4px;
   margin: auto;
-  padding: 3px;
+  padding: 4px;
   text-align: center;
   font-size: 12px;
 }
@@ -52,22 +49,57 @@
   color: white;
   margin: auto;
   text-align: center;
-  padding: 3px;
-  border-radius: 3px;
+  padding: 4px;
+  border-radius: 4px;
   font-size: 14px;
 }
 
 .join-btn {
   background-color: transparent;
-  border-radius: 3px;
-  border-color: grey;
+  border: 1px solid grey;
+  border-radius: 4px;
   margin: auto;
+  padding: 5px;
 }
 
 .leave-btn {
   margin: auto;
   color: red;
   background-color: white;
-  border-color: red;
-  border-radius: 3px;
+  border: 1px solid red;
+  border-radius: 4px;
+  padding: 5px;
+}
+
+.mission-profile {
+  margin-left: 2rem;
+  margin-top: 1rem;
+  width: 40%;
+}
+
+.mission-profile h2 {
+  margin-bottom: 1rem;
+}
+
+.mission-profile .no-missions {
+  border: none;
+  margin-top: 10px;
+}
+
+.mission-reservations {
+  border: 1px solid lightgrey;
+  padding: 20px 10px;
+}
+
+.member-sign {
+  padding: 10px;
+  border-right: solid 1px #d3d3d3;
+  display: flex;
+    align-items: center;
+}
+
+.btn-container {
+  padding: 10px;
+  display: flex;
+  align-items: center;
 }
