+++ src/components/MyProfile.jsx
@@ -1,7 +1,11 @@
-import React from "react";
+import MissionProfile from "./Mission/MissionProfile";
 
 const MyProfile = () => {
-  return <div></div>;
+  return (
+  <div className="mission-profile">
+    <MissionProfile />
+  </div>
+  );
 };
 
 export default MyProfile;
