+++ src/pages/Missions.js
@@ -1,30 +1,29 @@
-import Mission from '../components/Mission';
-import { useSelector, useDispatch } from 'react-redux';
-import { useEffect } from 'react';
-import { getMissions } from '../redux/mission/missionSlice';
+import Mission from '../components/Mission/Mission';
+import { useSelector } from 'react-redux';
 
 const Missions = () => {
-  const dispatch = useDispatch();
-  const { missionList, status } = useSelector((state) => state.missions);
-
-  useEffect(() => {
-    if (status === 'idle' || status === 'succeed') {
-      dispatch(getMissions());
-    }
-  }, [status, dispatch]);
+  const { missionList } = useSelector((state) => state.missions);
 
   return (
-    <div>
-      {missionList.map((mission) => (
-        <Mission
-          key={mission.id}
-          id={mission.id}
-          name={mission.name}
-          description={mission.description}
-          reserved={mission.reserved}
-        />
-      ))}
-    </div>
+    <>
+      <div className="missions-info">
+        <p>Mission</p>
+        <p>Description</p>
+        <p>Status</p>
+        <p></p>
+      </div>
+      <div>
+        {missionList.map((mission) => (
+          <Mission
+            key={mission.id}
+            id={mission.id}
+            name={mission.name}
+            description={mission.description}
+            reserved={mission.reserved}
+          />
+        ))}
+      </div>
+    </>
   );
 };
